fix: clear the ship-build warnings, drop the duplicate Replicated toggle
Warnings first. All of these were pre-existing and none changed behaviour:

  - AssetCache's PendingLoad/PendingPublish are poor-man's variants, some
    fields mesh-only and some material-only, so the four aggregate
    initializers left members unnamed. Naming them explicitly as empty is
    better than suppressing the warning: it distinguishes "belongs to the
    other variant" from "someone forgot a field", which is exactly what
    -Wmissing-field-initializers exists to ask.
  - PumpPublishes accumulated meshMs and matMs that nothing read, and
    flushMs that was (void)-suppressed for the same reason. The log line
    that consumed them became profile scopes, which give the same
    mesh/material/flush split scrubbable and nested under the frame — so
    all three are removed rather than left as half-suppressed dead
    arithmetic that would have to be kept honest.
  - NetIdOfRef had no callers. Its inverse ComponentIdOfRef does (the
    removal diff), which is why only one of the pair survives.
  - EditorNet's join timer passed a float to a %f, which varargs promotes
    anyway; the cast just says so.

The inspector's "Replicated" checkbox is gone. Presence of the marker
component is the gate and now the only gate — the marker is an ordinary
component, added and removed through the same Add Component menu and the
same undoable path as anything else, so a second control meaning the same
thing was only a second place for the two to disagree.

Explicit-width integers where the type is ours to choose: test loop
counters and the input-polling loops (whose bounds were already int32_t).
Left alone at C API boundaries, where the type is dictated and a change
would only add casts — waitpid status, sysconf's long, GNS's returned
count, argc, __cpuid's unsigned int, NVML, stb_image's int* out-params,
GLFW's int* framebuffer size, ImGui::InputInt and ::Combo index params.

13/13 ctest suites green; both debug and ship builds now compile with no
warnings at all.

+++ modules/Editor/src/EditorNet.cpp
@@ -841,7 +841,11 @@ void EditorApp::DrawNetworkWindow()
         LabelledValue("Clock corrections", std::format("{}", stats.clockCorrections));
 
         if (_joinPhase == JoinPhase::Connecting)
-            ImGui::TextColored(kWarnColor, "Joining — waiting for the host (%.0f s)...", _joinElapsed);
+            // Explicit widening: %f consumes a double through varargs, so the
+            // float would promote anyway — saying so silences -Wdouble-promotion
+            // without changing a byte of behaviour.
+            ImGui::TextColored(kWarnColor, "Joining — waiting for the host (%.0f s)...",
+                               static_cast<double>(_joinElapsed));
         else if (!stats.worldComplete)
             ImGui::TextColored(kWarnColor, "Still receiving the initial world...");
     }
